Move backend specific backtrace code to backtrace_prim.c
And share backtrace.c between asmrun and byterun.
