+++ asmrun/backtrace_prim.c
@@ -19,52 +19,28 @@
 
 #include "caml/alloc.h"
 #include "caml/backtrace.h"
+#include "caml/backtrace_prim.h"
 #include "caml/memory.h"
 #include "caml/misc.h"
 #include "caml/mlvalues.h"
 #include "stack.h"
-
-int caml_backtrace_active = 0;
-int caml_backtrace_pos = 0;
-code_t * caml_backtrace_buffer = NULL;
-value caml_backtrace_last_exn = Val_unit;
-#define BACKTRACE_BUFFER_SIZE 1024
 
 /* In order to prevent the GC from walking through the debug information
    (which have no headers), we transform frame_descr pointers into
    31/63 bits ocaml integers by shifting them by 1 to the right. We do
    not lose information as descr pointers are aligned.
 
-   In particular, we do not need to use [caml_initialize] when setting
+   In particular, we do not need to use [caml_modify] when setting
    an array element with such a value.
 */
-#define Val_Descrptr(descr) Val_long((uintnat)descr>>1)
-#define Descrptr_Val(v) ((frame_descr *) (Long_val(v)<<1))
-
-/* Start or stop the backtrace machinery */
-
-CAMLprim value caml_record_backtrace(value vflag)
+value caml_raw_backtrace_slot_of_code(code_t pc)
 {
-  int flag = Int_val(vflag);
-
-  if (flag != caml_backtrace_active) {
-    caml_backtrace_active = flag;
-    caml_backtrace_pos = 0;
-    if (flag) {
-      caml_backtrace_last_exn = Val_unit;
-      caml_register_global_root(&caml_backtrace_last_exn);
-    } else {
-      caml_remove_global_root(&caml_backtrace_last_exn);
-    }
-  }
-  return Val_unit;
+  return Val_long((uintnat)pc>>1);
 }
 
-/* Return the status of the backtrace machinery */
-
-CAMLprim value caml_backtrace_status(value vunit)
+code_t caml_raw_backtrace_slot_code(value v)
 {
-  return Val_bool(caml_backtrace_active);
+  return ((code_t)(Long_val(v)<<1));
 }
 
 /* returns the next frame descriptor (or NULL if none is available),
@@ -194,7 +170,7 @@ CAMLprim value caml_get_current_callstack(value max_frames_value) {
     for (trace_pos = 0; trace_pos < trace_size; trace_pos++) {
       frame_descr * descr = caml_next_frame_descriptor(&pc, &sp);
       Assert(descr != NULL);
-      Field(trace, trace_pos) = Val_Descrptr(descr);
+      Field(trace, trace_pos) = caml_raw_backtrace_slot_of_code((code_t) descr);
     }
   }
 
@@ -203,11 +179,11 @@ CAMLprim value caml_get_current_callstack(value max_frames_value) {
 
 /* Extract location information for the given frame descriptor */
 
-CAMLexport void extract_location_info(frame_descr * d,
-                                  /*out*/ struct caml_loc_info * li)
+void caml_extract_location_info(code_t slot, /*out*/ struct caml_loc_info * li)
 {
   uintnat infoptr;
   uint32_t info1, info2;
+  frame_descr * d = (frame_descr *)slot;
 
   /* If no debugging information available, print nothing.
      When everything is compiled with -g, this corresponds to
@@ -241,121 +217,6 @@ CAMLexport void extract_location_info(frame_descr * d,
   li->loc_endchr = ((info2 & 0xF) << 6) | (info1 >> 26);
 }
 
-/* Print location information -- same behavior as in Printexc
-
-   note that the test for compiler-inserted raises is slightly redundant:
-     (!li->loc_valid && li->loc_is_raise)
-   extract_location_info above guarantees that when li->loc_valid is
-   0, then li->loc_is_raise is always 1, so the latter test is
-   useless. We kept it to keep code identical to the byterun/
-   implementation. */
-
-static void print_location(struct caml_loc_info * li, int index)
-{
-  char * info;
-
-  /* Ignore compiler-inserted raise */
-  if (!li->loc_valid && li->loc_is_raise) return;
-
-  if (li->loc_is_raise) {
-    /* Initial raise if index == 0, re-raise otherwise */
-    if (index == 0)
-      info = "Raised at";
-    else
-      info = "Re-raised at";
-  } else {
-    if (index == 0)
-      info = "Raised by primitive operation at";
-    else
-      info = "Called from";
-  }
-  if (! li->loc_valid) {
-    fprintf(stderr, "%s unknown location\n", info);
-  } else {
-    fprintf (stderr, "%s file \"%s\", line %d, characters %d-%d\n",
-             info, li->loc_filename, li->loc_lnum,
-             li->loc_startchr, li->loc_endchr);
-  }
-}
-
-/* Print a backtrace */
-
-void caml_print_exception_backtrace(void)
-{
-  int i;
-  struct caml_loc_info li;
-
-  for (i = 0; i < caml_backtrace_pos; i++) {
-    extract_location_info((frame_descr *) (caml_backtrace_buffer[i]), &li);
-    print_location(&li, i);
-  }
-}
-
-/* Convert the raw backtrace to a data structure usable from OCaml */
-
-CAMLprim value caml_convert_raw_backtrace_slot(value backtrace_slot) {
-  CAMLparam1(backtrace_slot);
-  CAMLlocal2(p, fname);
-  struct caml_loc_info li;
-
-  extract_location_info(Descrptr_Val(backtrace_slot), &li);
-
-  if (li.loc_valid) {
-    fname = caml_copy_string(li.loc_filename);
-    p = caml_alloc_small(5, 0);
-    Field(p, 0) = Val_bool(li.loc_is_raise);
-    Field(p, 1) = fname;
-    Field(p, 2) = Val_int(li.loc_lnum);
-    Field(p, 3) = Val_int(li.loc_startchr);
-    Field(p, 4) = Val_int(li.loc_endchr);
-  } else {
-    p = caml_alloc_small(1, 1);
-    Field(p, 0) = Val_bool(li.loc_is_raise);
-  }
-
-  CAMLreturn(p);
-}
-
-/* Get a copy of the latest backtrace */
-
-CAMLprim value caml_get_exception_raw_backtrace(value unit)
-{
-  CAMLparam0();
-  CAMLlocal1(res);
-  const int tag = 0;
-
-  /* Beware: the allocations below may cause finalizers to be run, and another
-     backtrace---possibly of a different length---to be stashed (for example
-     if the finalizer raises then catches an exception).  We choose to ignore
-     any such finalizer backtraces and return the original one. */
-
-  if (caml_backtrace_buffer == NULL || caml_backtrace_pos == 0) {
-    res = caml_alloc(0, tag);
-  }
-  else {
-    code_t saved_caml_backtrace_buffer[BACKTRACE_BUFFER_SIZE];
-    int saved_caml_backtrace_pos;
-    intnat i;
-
-    saved_caml_backtrace_pos = caml_backtrace_pos;
-
-    if (saved_caml_backtrace_pos > BACKTRACE_BUFFER_SIZE) {
-      saved_caml_backtrace_pos = BACKTRACE_BUFFER_SIZE;
-    }
-
-    memcpy(saved_caml_backtrace_buffer, caml_backtrace_buffer,
-           saved_caml_backtrace_pos * sizeof(code_t));
-
-    res = caml_alloc(saved_caml_backtrace_pos, tag);
-    for (i = 0; i < saved_caml_backtrace_pos; i++) {
-      /* [Val_Descrptr] always returns an immediate. */
-      Field(res, i) = Val_Descrptr(saved_caml_backtrace_buffer[i]);
-    }
-  }
-
-  CAMLreturn(res);
-}
-
 CAMLprim value caml_add_debug_info(code_t start, value size, value events)
 {
   return Val_unit;
@@ -366,29 +227,7 @@ CAMLprim value caml_remove_debug_info(code_t start)
   return Val_unit;
 }
 
-/* the function below is deprecated: we previously returned directly
-   the OCaml-usable representation, instead of the raw backtrace as an
-   abstract type, but this has a large performance overhead if you
-   store a lot of backtraces and print only some of them.
-
-   It is not used by the Printexc library anymore, or anywhere else in
-   the compiler, but we have kept it in case some user still depends
-   on it as an external.
-*/
-
-CAMLprim value caml_get_exception_backtrace(value unit)
+int caml_debug_info_available(void)
 {
-  CAMLparam0();
-  CAMLlocal3(arr, res, backtrace);
-  intnat i;
-
-  backtrace = caml_get_exception_raw_backtrace(Val_unit);
-
-  arr = caml_alloc(Wosize_val(backtrace), 0);
-  for (i = 0; i < Wosize_val(backtrace); i++) {
-    Store_field(arr, i, caml_convert_raw_backtrace_slot(Field(backtrace, i)));
-  }
-
-  res = caml_alloc_small(1, 0); Field(res, 0) = arr; /* Some */
-  CAMLreturn(res);
+  return 1;
 }
