+++ asmrun/stack.h
@@ -78,15 +78,6 @@ typedef struct {
   unsigned short live_ofs[1];
 } frame_descr;
 
-struct caml_loc_info {
-  int loc_valid;
-  int loc_is_raise;
-  char * loc_filename;
-  int loc_lnum;
-  int loc_startchr;
-  int loc_endchr;
-};
-
 /* Hash table of frame descriptors */
 
 extern frame_descr ** caml_frame_descriptors;
@@ -99,10 +90,6 @@ extern void caml_init_frame_descriptors(void);
 extern void caml_register_frametable(intnat *);
 extern void caml_unregister_frametable(intnat *);
 extern void caml_register_dyn_global(void *);
-
-CAMLextern void extract_location_info(frame_descr * d,
-                                      /*out*/ struct caml_loc_info * li);
-
 
 extern uintnat caml_stack_usage (void);
 extern uintnat (*caml_stack_usage_hook)(void);
