+++ byterun/Makefile.common
@@ -20,7 +20,7 @@ CC=$(BYTECC)
 COMMONOBJS=\
   interp.o misc.o stacks.o fix_code.o startup_aux.o startup.o \
   freelist.o major_gc.o minor_gc.o memory.o alloc.o roots.o globroots.o \
-  fail.o signals.o signals_byt.o printexc.o backtrace.o \
+  fail.o signals.o signals_byt.o printexc.o backtrace_prim.o backtrace.o \
   compare.o ints.o floats.o str.o array.o io.o extern.o intern.o \
   hash.o sys.o meta.o parsing.o gc_ctrl.o terminfo.o md5.o obj.o \
   lexing.o callback.o debugger.o weak.o compact.o finalise.o custom.o \
@@ -30,7 +30,7 @@ PRIMS=\
   alloc.c array.c compare.c extern.c floats.c gc_ctrl.c hash.c \
   intern.c interp.c ints.c io.c lexing.c md5.c meta.c obj.c parsing.c \
   signals.c str.c sys.c terminfo.c callback.c weak.c finalise.c stacks.c \
-  dynlink.c backtrace.c
+  dynlink.c backtrace_prim.c backtrace.c
 
 PUBLIC_INCLUDES=\
   address_class.h alloc.h callback.h config.h custom.h fail.h gc.h \
